requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==5.1.1
Django==5.1.3
django-appconf==1.0.6
django-redis==5.4.0
django-select2==8.2.1
djangorestframework==3.15.2
executing==2.1.0
idna==3.10
ipython==8.29.0
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.2.3
wcwidth==0.2.13
